docs(example): Refactor examples for better readability
